renamed comopnent to page
