+++ src/App.js
@@ -2,7 +2,7 @@ import React from "react";
 import { HashRouter as Router, Route } from "react-router-dom";
 import "react-datepicker/dist/react-datepicker.css";
 import "./App.scss";
-import TransientTaxForm from "./pages/TransientTaxForm";
+import TransientTaxFormPage from "./pages/TransientTaxFormPage";
 import ConfirmationPage from "./pages/ConfirmationPage";
 import { Config } from "@baltimorecounty/javascript-utilities";
 import ErrorPage from "./pages/ErrorPage";
@@ -29,7 +29,7 @@ function App() {
   return (
     <div className="tt_app">
       <Router>
-        <Route exact path="/" component={TransientTaxForm} />
+        <Route exact path="/" component={TransientTaxFormPage} />
         <Route
           exact
           path="/ConfirmationPage/:confirmationNumber"
